pyproject.toml
[project]
name = "fakefinder-website"
version = "0.1.0"
description = "Détecteur de deepfakes avec Grad-CAM"
authors = [{name = "Julien Lucas", email = "hello@julienlucas.com"}]
requires-python = ">=3.11,<3.13"
dependencies = [
    "django>=5.2.6",
    "numpy<2.0.0",
    "pillow>=11.0.0",
    "python-dotenv>=1.1.1",
    "werkzeug>=1.0.1",
    "uvicorn>=0.24",
    "onnxruntime>=1.16.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "onnxscript>=0.5.7",
]
